requirements.txt
asyncio_mqtt==0.12.1
discord.py==1.7.3
lxml==4.8.0
PyPDF2==1.26.0
python-dotenv==0.19.2
requests==2.27.1
tika==1.24
discord-py-slash-command==3.0.3
aenum==3.1.8
mysql-connector-python==8.0.28
google-api-python-client==2.42.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1